requirements.txt
amqp==5.0.9
asgiref==3.4.1
attrs==21.4.0
billiard==3.6.4.0
blinker==1.4
celery==5.2.3
certifi==2021.10.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2
django-celery-beat==2.2.1
django-cors-headers==3.10.1
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
dnspython==2.2.0
drf-spectacular==0.21.1
drf-spectacular-sidecar==2022.1.24
gunicorn==20.1.0
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
jsonschema==4.4.0
kombu==5.2.3
mongoengine==0.23.1
packaging==21.3
pip-autoremove==0.10.0
pluggy==1.0.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.0
pymongo==3.12.3
pyparsing==3.0.6
pyrsistent==0.18.1
pytest==6.2.5
pytest-django==4.5.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.1.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sentry-sdk==1.5.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wait-for-it==2.2.1
wcwidth==0.2.5
wrapt==1.13.3
